requirements.txt
gitdb==4.0.11
GitPython==3.1.43
smmap==5.0.1